Modify the mixed login activity
Mixed Login Activity was extending LoginActivity which resulted in a
complicated test settings where two layouts are active at a time.
Separating the Mixed LoginActivity from the LoginActivity so that one
layout is active at a time. This might be the fix for a partner whose
tests were failing because the views were not registering click
properly.

Flag: TEST_ONLY
Test: atest CtsAutoFillServiceTestCases:android.autofillservice.cts.inline.InlineLoginMixedCredentialActivityTest
Bug: 340821203
Change-Id: I9b86ba2b7505c9c7e5863c2f3da68f114722fc81