Use Display#getRealMetrics of default display to create mirror display
The physical width, height of the default display was being used to
create the auto-mirror virtual display in VirtualDeviceMirroDisplayTest.
When the default display has scaling applied, this creates a mismatch of
bounds between the default display and the mirror virtual display. Using
Display#getRealMetrics fixes this issue as it takes the display scaling
into account.

Test: adb shell wm size 640x480 && atest VirtualDeviceMirrorDisplayTest
Flag: EXEMPT test fix
Bug: 357552750
Change-Id: I96a1531cbc5393e9c4dc7f0c2b6e5800665d0c57