Remove assumption of settings app handing panel actions.
There is no CDD requirement for quick settings panel actions to be
handled by the android settings app, as OEMs are free to provide
customized settings panels.

This CL fixes this false assumption in SettingsPanelTest by verifying
the package name of the panel shown matches that of the provider of the
action itself, instead of assuming it will always be the settings app.

The tests that verify opening the settings app are also modified to
assumption fail early if the handler for the corresponding settings
actions isn't the system app, as it doesn't have to point to the
system app.

Bug: 357918738
Test: Tests pass when panel UI is displayed by an override provider.
Change-Id: I5987cd199bea126c84af543fdd66c45bc0724246