Make music audio focus acquisition more resilient to failure.
When calling AudioManager#requestAudioFocus it is possible to get
AUDIOFOCUS_REQUEST_DELAYED as a return value.  This means that the
listener will get informed when the audio focus changes to "focused".

To make this more resilient, we will now try to get focus for up to 5
seconds, trying every 1/2 second if we get a focus failure.  If we get
a focus delay, we rely on the listener to report when we do get focus.

This makes us more resilient to timing delays or prior test failures.

Test: atest CtsTelecomCujTestCases
Flag: EXEMPT test change
Fixes: 357916523
Change-Id: I0695cb35051da9854507082ac36366d50bd8fb28