Fix DecodeOnlyTest race condition when releasing the audio codec
If a callback fires and the audio playback thread is writing to the
audio track, then the audio codec can be released during this time,
causing subsequent release of the audio codec buffer to throw an
exception.

Fixed by immediately releasing the audio codec buffer after copying the
data and before writing to the audio track.

Also cleaned up and simplified the code.

Bug: 351133404
Test: atest DecodeOnlyTest
Change-Id: I0bde3abe283610c8f5796509c76dab113c50f547