Use default documentLaunchMode for Landscape/PortraitOrientationActivity.
In certain freeform modes, activities launched with WINDOWING_MODE_FULLSCREEN may be launched in different freeform tasks and not the same tasks if documentLaunchMode is set to always.

Call launchFullscreenPortraitActivityToFront in test cases where activities are launched one after the other and should be on top in some order when checking for visibility.

Change-Id: I7c87717102011b27d99139d4e5034c9c2df9b658
Test: atest AppConfigurationTests
Bug: 358418625
Bug: 353335964
Flag: TEST_ONLY