Mark flaky test ExtendedInCallServiceTest test.
Marks the ExtendedInCallServiceTest#testSwitchAudioRoutes test as flaky.
This test is recording flakiness due to how the broadcasts are processed
by the state machine. Given a scenario where we switch in and then out
of speakerphone, it's possible for the SPEAKER_ON/SPEAKER_OFF requests
to be processed in reverse order. This is due to the fact that the
requests are queued up at the front of the queue so if the state machine
begins processing the after the SPEAKER_OFF message is received, it will
process that request first followed by the SPEAKER_ON message causing
the audio to incorrectly route back to speakerphone.

This is addressed in the call audio refactored changes where we place
the requests at the end of the message queue so that the order is
preserved. We can unmark this test as flaky once that flag has been
enabled.

Bug: 344112674
Test: atest ExtendedInCallServiceTest#testSwitchAudioRoutes
Flag: TEST_ONLY
Change-Id: I7e24c7971e0329efcff77a55c9c8ecfcd0bfb5b4