ITS: For preview req, set the intent to PREVIEW
Currently STILL_CAPTURE intent is being used by default for all the
requests in ItsSerializer#deserializeRequestList. Update the logic to
check the value of capture intent in json object and if the value is 1,
set the intent to PREVIEW for capture request and use PREVIEW capture
template else use STILL_CAPTURE.

Updated scene1_1/test_ae_precapture_trigger to set the capture intent
to PREVIEW for the precap requests.

Bug: 358407479
Test: Tested locally on pixel device.
Flag: TEST_ONLY
Change-Id: Ib93a12c60c5f8540d0c44a6471006ebc5a5d9bb4