Fixed CtsBiometricsTestCases test failures.
The test would fail when using am start with the -W option to
wait for the activity to launch, which calls
ActivityTaskManagerService#startActivityAndWait. This fix reverts to
the original approach used in ActivityManagerTestBase#launchActivity,
removing the -W option and waiting for the window state to be ready
instead, which calls ActivityTaskManagerService#startActivityAsUser.

Test: atest CtsBiometricsTestCases:BiometricActivityTests CtsBiometricsTestCases:BiometricPromptLogoTests
Bug: 360735368
Change-Id: I4ba591cf9c4fee56e0beee8788e61246c2f5e63b