Deflake WebViewSslTest#testClearSslPreferences attempt 2
WebViewSslTest#testClearSslPreferences was flaking occasionally as the
SSL error callback was not being invoked at the expected time.

WebView requests a favicon from sites after a page loads successfully,
which causes an additional request to be sent to the same host that the
test was not expecting. If this request comes after the test has cleared
the cached SSL preferences, it can disrupt the test by creating a new
cached entry sooner than expected.

Avoid this by waiting for the page's favicon to be set after the initial
load has completed.

Bug: 353493560
Bug: 361740568
Test: atest WebViewSslTest
Flag: TEST_ONLY
Change-Id: I9024ec7e462b66cb071a9a9ed3860819bbc350dd