Fix CTS tests that toggle DND
These tests toggle ringer mode and checks which streams are muted. The
set up for the test is incomplete as both the tests indirectly enable
DND causing the system default notification policy to be applied. The
test verifies certain streams are not muted, but that depends on the
system default notification policy. The fix correctly sets the policy
before verification

Flag: EXEMPT test fix
Bug: 360428132
Test: atest android.media.audio.cts.AudioManagerTest
Change-Id: I17860fea3292b11b39e7161faaeb4f7914590082