Pass --user cur flag to ItsService when running tests
In HSUM, user 0 is the system user and may not be the current user. This
will lead to a permission denial, as the system user is not the
foreground user and may not be on the list of allowed users. Run as the
current user instead.

Bug: 357733659
Test: Ran scene0 in hsum / non-hsum
Flag: EXEMPT its python code
Change-Id: Ie7afbe3b4abee356d1c594b6e0a2f0a588ec0842