DO NOT MERGE Remove unnecessary and error prone setup methods.
Setup methods tries to inject events to wake up device
and unlock it which is not related to the tests and
the device should be in a ready state. Instead these
injected events pose problems as they are sometimes
injected before the test activity has started causing
a KeyDispatchingTimeout and test execution stops.

Test: atest CtsWindowManagerDeviceTestCases
:android.server.wm.WindowUntrustedTouchTest

Bug: 362511273
Change-Id: I69938ad807e677950488820d7cf9234381a4da03