Enable secondary_user_on_secondary_display on CtsRenderscriptTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices
that have config_multiuserVisibleBackgroundUsers enabled. The main use
case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) interacting with driver's display.

Bug: 363126761
Flag: TEST_ONLY
Test: atest -c --user-type secondary_user_on_secondary_display CtsRenderscriptTestCases
Change-Id: Ia1c77621b1a050a131c2b9bcf8cac9ee70c2a27d