Enable secondary_user_on_secondary_display for CtsVideoCodecTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled. The
main use case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the
display in front of them concurrently with the driver (modeled as the
current user) interacting with driver's display.

Bug: 364450820
Test: atest CtsVideoCodecTestCases
Test: atest --user-type secondary_user_on_secondary_display \
        CtsVideoCodecTestCases

Change-Id: Icd9f114cdda678e19d642222fd2a7b674ce7744f