Resolve flaky SelfManagedConnectionServiceTest test
SelfManagedConnectionServiceTest#isInSelfManagedCall has recently
started showing a high flake rate in main for aosp cf targets. After
looking at the logs of the failed test runs, it looks like the self
managed call will disconnect if it's in the midst of removing an
existing call. This test can be improved so that we only create one self
managed call and reuse it to verify the different cases presented by
this test case.

Bug: 361490107
Test: SelfManagedConnectionServiceTest
Flag: TEST_ONLY
Change-Id: I1fd29aa4d785f23c389be9e2c5f907d0f84f9205