Ensure current user is used when setting the PhoneAccountSuggestionService.
This CL updates the relevant CTS TestUtils method to ensure the current primary user is being used when setting the PhoneAccountSuggestionService for Telecom CTS test purposes. In HSUM mode, this should be user 10 and in non HSUM mode this should be user 0. Without this change, PhoneAccountSuggestionServiceTest tests were failing in HSUM mode.

Test: atest PhoneAccountSuggestionServiceTest
Bug: 358587742
Flag: TEST_ONLY
Change-Id: I46d7a9bd0dc8a3bac1cfa799c7575e855d39b317