Enable secondary_user_on_secondary_display for CtsCredentialManagerTestCases and CtsResourcesTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled.
The main use case is Automotive's multi-display Whole Cabin experience
where passengers (modeled as visible background users) can interact
with the display in front of them concurrently with the driver
(modeled as the the current user) interacting with driver's display.

Bug: 364413340
Test: atest CtsCredentialManagerTestCases && atest CtsResourcesTestCases
Test: also tested --user-type secondary_user and --user-type secondary_user_on_secondary_display
Flag: TEST_ONLY

Change-Id: I4ca2746a2867e713534ee21ae5e23a4298252e30