Telecom: drop TestUtils.waitForCount(timeout)
This method is flake-prone: it's equivalent of taking current call count
and waiting for it to increase by 1. But in all use cases it happens after
trigger (callback setup) has already happened, so we're waiting for the
other thread to catch up (which might have already happened).

This also fixes IndexOutOfBoundsException if timeout hits without satisfying
counter tick condition.

Bug: 365146607
Test: atest CtsTelecomTestCases:android.telecom.cts.ExtendedInCallServiceTest#testSwitchCallEndpoint
      on device without telephony
Flag: TEST_ONLY
Change-Id: I4fef361ee0aff3024a790665216bc41927435e8f