Extend the test for the LRU order of isolated connections.
The test now also verifies that the LRU order of isolated connections
that do not have any importance group set is correct, both for strong
and weak connections.

Binding isolated processes from a client process
in the order of 1,2,3,4,5 was putting the processes
in the LRU list in this order: [LRU]4,3,2,1,5,client[MRU].
But it actually needs to be [LRU]1,2,3,4,5,client[MRU].

This CL extends the existing test to cover this case as well.

Bug: 356397711
Change-Id: Ie886ef587b1ecc5d8814a9fa63501781aa6ddb6e
Test: atest CtsAppTestCases:ServiceTest#testActivityServiceBindingLru
Flag: EXEMPT bugfix