Add fallback to previous resource name for MediaProjection tests
The spinner for the MediaProjection dialog was called screen_share_mode_spinner, but was renamed to screen_share_mode_options in ag/28231322. Now builds from 24Q3 or earlier use the first resource id while builds from 24Q4 onwards use the second. To ensure that all tests work with all builds, we are updating tests to check both the resource ids when referring to the spinner.

Bug: 362402351
Bug: 361027310
Bug: 362189183
Bug: 363868335
Test: atest CtsMediaProjectionTestCases
Test: atest CtsMediaProjectionSDK33TestCases
Test: atest CtsMediaProjectionSDK34TestCases
Flag: TEST_ONLY
Change-Id: If34e3500b4c78681fa94dce887f5516e263c9598