Skip tests that use camera for visible background users.
Camera is not supported for visible background users, therefore skip
the tests in MediaPlayerTest that use camera when running with
--user-type secondary_user_on_secondary_display.

secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled. The
main use case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the
display in front of them concurrently with the driver (modeled as the
current user) interacting with driver's display.

Bug: 336880935

Test: atest CtsMediaPlayerTestCases
Test: atest --user-type secondary_user_on_secondary_display \
        CtsMediaPlayerTestCases
Change-Id: I51ee4412fda32b48f178da57272be1ad7cf1f993