Fix CtsContactsProviderWipe for the visible background user
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices that have config_multiuserVisibleBackgroundUsers enabled. The main use case is Automotive's multi-display Whole Cabin experience where passengers (modeled as visible background users) can interact with the display in front of them concurrently with the driver (modeled as the the current user) interacting with driver's display.

To wipe the contacts provider, the pm clear command is used. The user
option value of the command is always set to current, so it fails when
tested with the visible background user. It has been modified to specify
the user option value to the testing userId, so the test can be passed
with the visible background user.

Bug: 365604982
Flag: TEST_ONLY
Test: atest CtsContactsProviderWipe
Test: atest --user-type secondary_user CtsContactsProviderWipe
Test: atest --user-type secondary_user_on_secondary_display CtsContactsProviderWipe
(cherry picked from https://partner-android-review.googlesource.com/q/commit:c8ce4cc8ee061c3645bf100120d3ba3481711f32)

Change-Id: I9a460ce60e040a8ca9232e1b0f6cea872e556cc2