Skip volume adjustment test for STREAM_VOICE_CALL.
Currently, the volume adjustment for STREAM_VOICE_CALL will be dynamical
based on whether a sco device is active or not. In that case, the step
for changing volume of STREAM_VOICE_CALL is not guaranteed to be 1.

In this CL, the volume adjustment test for STREAM_VOICE_CALL is skipped.
The test will be reenabled when there are APIs to query the ratio
between the voice call range transmitted to APM and the range of the
min/max stream volume returned by the AM API.

Bug: 362324888
Test: run the test
Flag: EXEMPT bugfix
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:bbe371da8c09633e61ff70b671118f7b1db6e350)
Merged-In: I54de301c06fbd2191f4d6536754f7ed26ed698b9
Change-Id: I54de301c06fbd2191f4d6536754f7ed26ed698b9