Check if it can start the stream when it is opened without permission.
When opening a mmap stream without right permission may not fail. The
reason is that it can fallback to open a shared stream, which is
connected to an exclusive stream opened by aaudio service. Note that the
shared stream without permission will not be able to start. In that
case, in the CTS test, it should check if the stream fails to start if
it successfully opens a mmap stream.

In this CL, it also added sharing mode as an input parameter for the
test. In that case, it can make sure the test pass with all valid
sharing mode.

Bug: 365782831
Test: atest AAudioTests
Flag: EXEMPT bugfix
Change-Id: Ib77627657a0592b69de3006d838bc8bca7c39265