Improve task cleanup in LocaleManager tests.
This change adds stopTestPackage() calls in the teardown methods of
LocaleManagerOverrideLocaleConfigTest and LocaleManagerTests to ensure
proper cleanup of test activities from test apps.

This change is necessary because ag/29358865 introduced a new way to
clean up tasks that is more targeted and only affects the current test
package.

Bug: 355452977
Flag: TEST_ONLY
Test: atest CtsLocaleManagerTestCases
Change-Id: Ie33a898ba1f671e31968168bab2ce6733b537770