Improve task cleanup in LocaleManager tests.
This change adds stopTestPackage() calls in the teardown methods of
LocaleManagerOverrideLocaleConfigTest and LocaleManagerTests to ensure
proper cleanup of test activities from test apps.

This change is necessary because ag/29358865 introduced a new way to
clean up tasks that is more targeted and only affects the current test
package.

This patch has been ported from the android15-qpr1-release branch because,
without this cleanup, the tests testGetApplicationLocales_nonCurrentIme_noBroadcast
and testGetApplicationLocales_currentImeGetForegroundAppLocales_returnsLocales
fail when executed as part of the CTS suite.

Bug: 404160268
Flag: TEST_ONLY
Test: atest CtsLocaleManagerTestCases
Change-Id: Ie33a898ba1f671e31968168bab2ce6733b537770
Signed-off-by: Salini Venate <salini.venate@intel.com>