Improve task cleanup in CtsDreamsTestCases tests.
This change adds stopTestPackage() calls in the teardown methods of
DreamOverlayTest and DreamServiceTest to ensure proper cleanup of test
activities from test apps.

This change is necessary because ag/29358865 introduced a new way to
clean up tasks that is more targeted and only affects the current test
package.

Bug: 355452977
Flag: TEST_ONLY
Test: atest CtsDreamsTestCases
Change-Id: I0716ba5263a5241135e883377f6503cbdf8f6ff1