PointerIconTest: Use VDM to create display to disable acceleration
When creating a virtual display using VDM, mouse acceleration is
automatically disabled. There is currently no other API accessible by
CTS that can be used to disable mouse acceleration for a display.

We are considering several appraoches to work around this in the longer
term, such as exposing these settings as display flags passed into
DisplayManager during display construction, or exposing APIs like
setMousePointerAccelerationEnabled via InputManager for testing.

In the meantime, we work around the issue to unblock the addition of new
tests for touchpads by using VDM to create virutal displays for
PointerIconTest, on which mouse acceleration will be automatically
disabled.

Bug: 332973324
Test: atest PointerIconTest
Flag: EXEMPT refactor
Change-Id: I5b93aca9ca7748ac67867aa850b6d0990010d110