DO NOT MERGE ActivityTransitionTests does not respect animation scale.
The test tries to determine acceptable duration of Activity transition
animation where it needs to within a certain time window. This window
does not respect the current animation scale and if not set to 1 (default)
they will always fail. The test also lacks one assert if the animation
never finishes which instead gives a confusing result.

Test: atest CtsWindowManagerDeviceTestCases:
android.server.wm.ActivityTransitionTests

Bug: 366410783
Change-Id: Ieca6088e8bd2c85d21e9585a5cd1fed566745901