Fix HotwordDetectionServiceBasicTest for Wear emulators
The @RequiresDevice annotation used to skip CTS on emulators does not
work for Wear emulators, so this change manually skips the annotated
tests for Wear emulators. b/367437694 will track a long term solution to
fix the annotation for this and all other tests that rely on the
annotation.

Bug: 365360680
Test: atest HotwordDetectionServiceBasicTest
Flag: TEST_ONLY
Change-Id: I2f10bf5e7a454e313aebd34c55bf6fb44b4e925c