Remove testSupportedDimensions()
testSupportedDimensions() was written with a wrong assumption that
media_codecs.xml must list the supported dimensions for all the video
codecs and not doing so was wrongly considered as an error.

When a codec doesn't list supported sizes, MediaCodecInfo treats this
as a codec that is capable of supporting all the dimensions and that
is an allowed behavior.

So the getSupportedWidths() and getSupportedHeight() in the codec
capabilities always returns a valid range even when the codec doesn't
advertise any supported sizes.

testSupportedDimensions() doesn't really achieve what it was meant to
and hence it is being removed now.

Bug: 314246256
Test: Builds

Change-Id: I422dc001fb90ddbcc70163fdfb1e155bf51702b7