Process CTS blocked number requests directly from BlockedNumberUtil.
This CL updates BlockedNumberUtil in order to process requests to insert and delete numbers in the BlockedNumberProvider directly from BlockedNumberUtil rather than binding to the BlockedNumberService and then processing those requests indirectly through that service. This is a neccessary change to support HSUM in Telecom since the BlockedNumberServioce is initialized within the context of user 0 but in HSUM it is neccessary to access the BlockedNumberProvider in order to insert and delete blocked numbers within the context of user 10 (the main user in HSUM mode).

Test: ExtendedInCallServiceTest#testIncomingCallFromBlockedNumber_IsRejected
Flag: TEST_ONLY
Bug: 358587742
Change-Id: Iee270e67c6269f08828cb40e49e71490b37f7b02