Remove errant library name from suites.
I think the original intent was to list `androidx.test.rules` in static_libs, not in test_suites.

I serves no purpose in test_suites and breaks some changes I'm making to test_suites, so I propose we remove it.
I'd add it to `static_libs`, but this test doesn't seem to need it there as it builds okay.  Some other dep must pull it in.

Change-Id: I26dd57e8215fbfcf99cf70bbccedfb1050af1af3
Test: just CI