Invoke invalidate to make a view get drawn in ViewTest
Calling requestLayout is not enough to make a view get re-drawn if
there is no dirty region and the full re-draw is not needed.

The previous logic of the test relied on the invocation of
ViewRootImpl#dispatchFocusEvent which would set mFullRedrawNeeded to
true. However, if dispatchFocusEvent is not called, or a traversal is
scheduled right after dispatchFocusEvent, the test would fail.

This CL calls View#invalidate instead to make the test more robust.

Bug: 337018949
Bug: 349601805
Flag: EXEMPT bugfix
Test: atest ViewTest#testVelocityAPIsFlagEnabled
Change-Id: I88296845ddc8f93c0c821260af2909f275161bb0