Correct CallDiagnosticServiceTest cleanup issue.
CallDiagnosticServiceTest was not cleaning up the phone account it
registered in the setup method.  This was causing issues with some of
the SelfManagedConnectionService tests due to the fact that the default
sim account is also a connection manager.

Also, cleaned up CallDetailsTest, ConnectionServiceTest, OutgoingCallTest,
and TransactionalApisTest, all which registered phone accounts and didn't
unregister them.

Also, added a proactive cleanup of all phone accounts after each test
runs. This helps ensure that poor test isolation issues are not going to
cause this kind of problem in the future (this found a test isolation
issue with the transactional tests).  If proactive cleanup happens, we will
fail a test to ensure we catch it in presubmit.

Flag: NONE - test only fix
Test: This is a test fix; ran the CtsTelecomTests locally in full to verify
they work as expected.
Fixes: 365199780
Fixes: 367864266

Change-Id: I3a4f9e9c7e8b7797b4ac8fce91fd97a7ea7a40ad