UinputTouchDevice: Account for display scaling
Account for cases where size of the logical display is different
that that of the physical display.

This is a partial solution that is temporary, and only works for cases
where the physical and logical displays use the same aspect ratios.

The full solution has landed internally, and requires platform
TestApi changes.

Bug: 365876900
Change-Id: Ie6a1ee0f0a747f9685d526451251b908301e0f3a
Test: Presubmit
Flag: TEST_ONLY
Merged-In: I2a597d1bcbb988a80a7bb49ce7814bb596a42f05