Remove `CtsJvmtiRunTest988HostTestCases` from CTS and ART MCTS.
Test module `CtsJvmtiRunTest988HostTestCases`, based on ART run-test
`988-method-trace`, uses the boot classpath (e.g. `java.lang.String`)
as a test fixture to exercise method tracing. This means that the test
depends on code that is unrelated to its concern (testing method
tracing), which may require updates when the boot classpath is changed
(as in
https://android-review.git.corp.google.com/q/topic:%22renameStringLatin1%22),
possibly leading to test failures (such as b/357156783, b/367976843,
or b/359777052). Remove this test from CTS and (ART) MCTS for now, in
order to avoid test issues with devices running Android 15 not
containing the above-mentioned internal behavioral changes to the boot
classpath.

Test: m cts
Bug: 357156783
Bug: 367976843
Bug: 359777052
Change-Id: If357ab36a1c1328019fba7543d89cfa83d610f7d