Strip out the user ID when checking temp power exemption list in MissedCallTest.
This CL updates the method that is responsible for confirming that the current uid is in the list of temporary power exemptions in MissedCallTest. This CL updates this method to strip out the user id in irder to just ensure that the app ID is present in that exemption list. This is more resilent in headless system user mode and fixes an HSUM CTS test failure.

Test: atest MissedCallTest
Bug: 358587742
Flag: TEST_ONLY
Change-Id: If9cceb9b29430c851c870af4b255bbe9c560cd75