Disable ColorSpaceTest on Ravenwood on AOSP
This test doesn't pass on AOSP yet.
Not sure why it was _not_ failing before,
but it might have been masked by the "tradefed silengly ignore
failures" issue.

We don't want this change in goog/main, so it has a "merged-in"
with an arbitrary change ID that's in goog/main.

Flag: EXEMPT host test change only
Bug: 292141694
Test: $ANDROID_BUILD_TOP/frameworks/base/ravenwood/scripts/run-ravenwood-tests.sh

Merged-in: Ied1e63416bbbc1986630a91856acf10a277924ae
Change-Id: I727fef84f46f7f11c5e2ca6078f775fc68e88c0a