Av1FilmGrainValidationTest: Remove first sdk check
When Av1FilmGrainValidationTest was part of mediav2 CTS tests,
it was limited to V launching devices. Now that it is moved to
the mediapc tests and only the devices that advertise media
performance class V are required by the test to handle film grain
correctly, the first sdk check is incorrect and hence removed.

Bug: 345354568
Test: atest android.mediapc.cts.Av1FilmGrainValidationTest

Change-Id: I6460f7ca0dc99c87f9e487f49fe3c565b257fe24