Disable ASM_RESTRICTIONS flag
This flag was added in Android 15, but introduces unexpected compatibility issues.
Grace period restriction were found impacting async launched activities, see go/asm-graceperiod-problems for details.

Impacted tests are ignored as we can't determine which state the device is in. 
This is a revert to Android 14 behaviour.

Bug: 367702727
Change-Id: I6d1ac2085694d177adff4c3a5a587ebb78573076
Flag: NONE Annotation change can't be flagged
Test: atest ActivitySecurityModelTest ActivitySecurityModelEmbeddingTest BackgroundActivityLaunchTest