CtsMediaEditing: Fix TranscodeQualityTest
TranscodeQualityTest is failing on meida3 version 1.4.0 because
trancoded ouput is not matching with input clip by 90%.
Issue is coming due to "Work around Surface Texture implicit scale"
Cl which crop the image by 2 pixels from each dimensions.
So updated TranscodeQualityTest to use
setExperimentalAdjustSurfaceTextureTransformationMatrix() which
restrict the cropping of the image.

NOTE: Issue is fixed in media3 version 1.5.0 so this change will
be removed once media3 1.5.0 is merged in aosp branch.

Bug: 357766120
Test: taest TranscodeQualityTest
Change-Id: I33e0f1c06aa1d81308934a0ddf5ef014aced7907