Check buffer before reading in PixelValidator
PixelValidator validates the number of correct pixels by
checking the pixels in a screen buffer one by one.  In
the event where the buffer was not filled completely or
correctly, a blind read can result in a buffer underflow
and crash the whole test instrumentation process, causing
all following scheduled tests to fail.

This CL mitigates this by first checking the buffer size
and only reading what's available.

Bug: 363026632
Test: No-op to passing tests.  Buffer underflows no longer crash test
instrumentation process.

Change-Id: I31f27605524154cc0d42d30580b4009ac1d26e60