Add tests for new KeyStoreManager grant APIs
Android 16 introduced new APIs to allow an app to grant access
to a key it owns in the Android KeyStore to another app on the
device. This commit adds CTS tests to verify these new APIs
function as expected. These tests are driven from a host side
test with a test app that owns the keys and controls granting /
revoking access to another app; the grantee app exports a bound
service with test methods that are invoked by the test app to
ensure access to the keys functions as expected based on whether
access has been granted or revoked.

Bug: 351158708
Test: atest KeyStoreManagerTest
Change-Id: Id7e1e44a1ba0669bea1e5c91adcd3c1d3166e9ca