Adjusting test case to not rely on receiving new insets, but being the same as before
With the flag turned on: When coming back from another app, the previous activity won't get new insets (as they're unchanged).
In this case, the test should check if the IME is visible and has the same insets as when going to the other app.

Test: atest KeyboardVisibilityControlTest#testImeVisibilityWhenImeTransitionBetweenActivities
Flag: android.view.inputmethod.refactor_insets_controller
Bug: 298172246
Change-Id: Iecd50a7652b0618c1f454a094424349487ad79aa