Add sleep between scrollUntil and click
This adds a short sleep in tests that use UiAutomator to scroll to a
view, and then click on that view. The issue is that UiAutomator eagerly
returns from the scrollUntil, before the scroll animation is finished,
and tries to click on the object. Because the animation is still
ongoing, the click won't be triggered, and the test will fail.

This can be triggered in the switchIme test by making sure there are
enough items in the IME Switcher Menu to require scrolling during the
test.

This should be a short term solution until the actual issue is fixed in
UiAutomator.

Flag: EXEMPT testfix
Bug: 371520375
Test: atest InputMethodManagerTest#testInputMethodPickerSwitchIme
  InputMethodManagerTest#testInputMethodPickerOpenLanguageSettings
Change-Id: Icbae7e606f53eda59547165274236c35bef18e89