Support secondary_user_on_secondary_display in some part of CtsWindowManagerDeviceDisplay
Some part of CtsWindowManagerDeviceDisplay test
always launches an activity on the default display and checks
the results there, even if the tests are performed on
the secondary user assigned to the secondary display.
To fix this, modify the logics to launch the activity and check
the result in the display assigned to the user.

Flag: EXEMPT (bug fix)
Bug: 358488239
Test: atest CtsWindowManagerDeviceDisplay:android.server.wm.display.AppConfigurationTests
Test: atest CtsWindowManagerDeviceDisplay:android.server.wm.display.AppConfigurationTests --user-type secondary_user
Test: atest CtsWindowManagerDeviceDisplay:android.server.wm.display.AppConfigurationTests --user-type secondary_user_on_secondary_display
Test: atest CtsWindowManagerDeviceDisplay:android.server.wm.display.PresentationTest
Test: atest CtsWindowManagerDeviceDisplay:android.server.wm.display.PresentationTest --user-type secondary_user
Test: atest CtsWindowManagerDeviceDisplay:android.server.wm.display.PresentationTest --user-type secondary_user_on_secondary_display
(cherry picked from https://partner-android-review.googlesource.com/q/commit:9b71040cc1a41e11186d9c940f0b61c5f36dd72a)
Change-Id: Ie8fa703357318dca20c880ec415067958816e24a