Resize CtsVerifier car tests buttons
Buttons in the car power policy test and garage mode test were too
large, causing them to be cut off or completely off the screen. By
adjusting the button heights, text font size, and margins, all of the
buttons are able to fit on the screen nicely.

FLAG: EXEMPT TEST-ONLY
Bug: 370415889
Test: Open Car Power Policy Test in CtsVerifier, 8 buttons are visible
Test: Open Garage Mode Test in CtsVerifier, 2 buttons are fully visible
Change-Id: Idd93838f8431de6846ee19a34f7485ba6feee133