Skip call notification tests for visible background users.
Call notification is not supported for visible background users,
therefore there is no need to run the test that expects media playback
to be suppressed by a call notification, with --user-type
secondary_user_on_secondary_display.

secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled. The
main use case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the
display in front of them concurrently with the driver (modeled as the
current user) interacting with driver's display.

Bug: 342432386

Test: atest CtsMediaCUJSmallTest
Test: atest --user-type secondary_user_on_secondary_display \
        CtsMediaCUJSmallTest

Change-Id: I361525f55c53d0bd4237c03d723d072aae493ca1