Extend ServiceTest#testActivityServiceBindingLru to include app resume
When a multiprocess app is sent to background and brought back, the
relative positions of its processes in the process LRU list was being
scrambled due to the connections being stored in an ArraySet, based on
the object's hash value, which is random. With ag/29464889,
updateLruProcessInternalLSP method is split into two stages
(offerLruProcessInternalLSP and completeLruProcessInternalLSP) having
a priority queue in between that stores and sorts the lrui (lru index)
to preserve the positions of the processes in the LRU list relative to
each other. Otherwise, when there is memory pressure, LMKD (Low Memory
Killer Daemon) was killing these processes in an undesired order.

This CL extends the existing test to include this case.

Bug: 359930209
Change-Id: Ic9189da01cbf474f3eb77a6b0c57acfd0e812a1b
Test: atest CtsAppTestCases:ServiceTest#testActivityServiceBindingLru
Flag: EXEMPT bugfix