Enable secondary_user_on_secondary_display for CtsContentTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled.
The main use case is Automotive's multi-display Whole Cabin experience
where passengers (modeled as visible background users) can interact
with the display in front of them concurrently with the driver
(modeled as the the current user) interacting with driver's display.

Fixes: 338132823
Test: atest CtsContentTestCases
Test: atest --user-type secondary_user CtsContentTestCases
Test: atest --user-type secondary_user_on_secondary_display CtsContentTestCases
Flag: TEST_ONLY

Change-Id: If8553642159f6c59b03173f3d89d79b7aba26a29