Fix failed testImeWindowCanSwitchToDifferentDisplays() for secondary_user_on_secondary_display
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled.
The main use case is Automotive's multi-display Whole Cabin experience
where passengers (modeled as visible background users) can interact
with the display in front of them concurrently with the driver
(modeled as the the current user) interacting with driver's display.

Bug: 355529421
Test: atest CtsWindowManagerDeviceIme
Test: atest --user-type secondary_user CtsWindowManagerDeviceIme
Test: atest --user-type secondary_user_on_secondary_display CtsWindowManagerDeviceIme
Flag: EXEMPT bugfix
Change-Id: Ie09e73ed0567fce669f02233d1f99e5eb498e7c2