Enable secondary_user_on_secondary_display for CtsFgsTimeoutTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled.
The main use case is Automotive's multi-display Whole Cabin experience
where passengers (modeled as visible background users) can interact
with the display in front of them concurrently with the driver
(modeled as the the current user) interacting with driver's display.

Fixes: 364646661
Test: atest CtsFgsTimeoutTestCases
Test: atest --user-type secondary_user CtsFgsTimeoutTestCases
Test: atest --user-type secondary_user_on_secondary_display CtsFgsTimeoutTestCases
Flag: TEST_ONLY

Change-Id: I4b127452bbb240516290c542d1793d60a3cd05be