[cts][5/N] use shell command to set policy
The test itself can no longer change the policy because it is not a
verifier. Use the new shell command to set the policy instead.

Also changed a test behavior slightly because session APIs cannot
be called by a non-verifier. A follow-up CL will re-implment the
original test on the verifier side.

FLAG: android.content.pm.verification_service

BUG: 360129103
BUG: 360129657
BUG: 367776952

Test: atest CtsPackageManagerTestCases:VerifierServiceTest

Change-Id: Icc060fbc20e6d1617d1cbb9afbc52a053630e04f