Make sure testRenameCanRestoreDeletedRowId is cleaning up created files
testRenameCanRestoreDeletedRowId could leave stale files in the database after
it finishes execution if race condition happens between file on close
listener and file deletion. This could lead to create a DB record in the
MediaProvider DB but delete the file from FS. And when this happens it
fails testPendingFromFuse, testListPendingAndTrashed and
testSystemGalleryCanUpdateOtherAndroidMediaFiles tests as they won't be
able to create the same same file again.

Added sleep time in testRenameCanRestoreDeletedRowId before trying to delete the created files to make sure that such race condition doesn't happen.

Bug: 360100744
Test: atest ScopedStorageDeviceTest
Flag: EXEMPT test fix
Change-Id: I6f239273621717adafe3930d7451669cd629365b