Fix AdvancedProtectionManagerTest#tearDown
In ag/29996609, we added a fix to skip tests in AdvancedProtectionManagerTest for form-factors that do not have AdvancedProtectionManager. This caused an issue in the tearDown() method; since the AdvancedProtectionManager is absent, any interaction with the mManager was causing NPEs. This change fixes this issue by skipping the tearDown() method altogether if no AdvancedProtectionManager has not been initialized.

Bug: 375055360
Change-Id: I17003e32e82d494c2f494764bbe59f7da9b231b6
Test: atest AdvancedProtectionManagerTest
Flag: TEST_ONLY