Fix testRequestBluetoothPermission_Downgrade by enabling location
This patch is from the commit "Enable location in NearbyDevicesPermissionTest" in AOSP main branch.
(Ifcba305a84402148283b9f6ecf96a66782f4b6db)

Location services must be enabled on the device to perform unfiltered
Bluetooth scans.

As a result, since this test isn't enabling location, it was at the
mercy of whether or not a previous test enabled/disabled it. The result
was that the test appeared flaky.

Bug: 347632842
Merged-In: If8856b546a8d4a733e3ec0139798b4495e604054
Change-Id: I28d93acf5620b99754078359322eec3af14d6f11