RESTRICT AUTOMERGE Also exclude "instant" FrameRateOverrideTest, DisplayTest from Android 15 CTS
The "@TestApi" for isSynthetic() did not make it for Android 15 24Q3 so FrameRateOverrideTest and some of the DisplayTest that uses isSynthetic() will be excluded in Android 15 CTS, but automatically included in Android 16 25Q2. A previous CL aosp/3324544 excluded the tests but the `instant` ones also need exclusion.

Bug: 370790791
Change-Id: I5de7b1347362fd274cb943b4e8370188a2d668a6
Test: presubmit
Flag: EXEMPT TEST_ONLY