Enable secondary_user_on_secondary_display for CtsDownloadManagerApi28
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled.
The main use case is Automotive's multi-display Whole Cabin experience
where passengers (modeled as visible background users) can interact
with the display in front of them concurrently with the driver
(modeled as the the current user) interacting with driver's display.

Fixes: 376743032
Test: atest CtsDownloadManagerApi28
Test: atest --user-type secondary_user CtsDownloadManagerApi28
Test: atest --user-type secondary_user_on_secondary_display CtsDownloadManagerApi28
Flag: TEST_ONLY
Change-Id: Icc8eaca06083304eea7bda36f2a1b4cf0371c5cc