RESTRICT AUTOMERGE Adds tests around addSearchEmbeddings and related to known failures
The fix for this cannot be cherrypicked (aosp/3308785) as new APIs were
also added in the same topic as the fix.

This change is for the run-on-work-profile configuration

Note that the Merged-In change is not this change, but another change already in aosp_main. We don't want this change in main

Test: CtsAppSearchTestCases
Bug: 367914183
Change-Id: Ia62f220b09fd22d35764f6f01505cf97a0617a76
Merged-In: I8b6c7c0a77b501d9c4687e76e6bb46a9e853c3b3