CameraPermissionTest: Test NDK path and fix proxy tests.
The NDK behaves slightly differently from Camera2/Camera1, in that
there is no Context to obtain an AttributionSource from. This deserves
its own set of tests.

Like Camera1, the NDK will not support AttributionSource chaining until
the API is modified to accept an AttributionSource.

On the other hand, the proxy tests had incorrect expectations. When
OpenCameraActivity enters the background, it should be blocked even if
it's included in the AttributionSource chain. This is unless it exports
a service which is bound with BIND_INCLUDE_CAPABILITIES. Refactor the
code that opens the camera to its own class, and export a service which
CameraProxyActivity will bind to in order to open the camera.

Bug: 190657833
Bug: 378016494
Bug: 379554373
Bug: 379515460
Flag: EXEMPT testing existing functionality
Test: Ran new tests on cuttlefish.
Change-Id: I90088adabe0fe9667bc17ce77407f825b46f6540