Remove test failing after bug fix
The test was testing that the virtual camera returns a black frame if the
producer does not write anything to the input surface. That behavior was
wrong and the test should not have been there.

Bug: 379232306
Bug: 364268484
Test: removed
Flag: TEST_ONLY
Change-Id: Ia173ddaae514ce6d8e200b3adf321d0df8d7855a
Merged-In: I833e1ed56969eda639bc9e5e8b98a5a7cf575634