bt(snippet): Add methods to OOB bond to the snippet.
This method will scan for the peer device advertising the specified UUID,
connect to it and then bond to it. The method will block until the
bonding is complete or fails.

This method will be used in the multi-device ranging tests to bond the
two devices before starting Channel sounding tests.

Also,
1. Added a method to remove the bond.
2. Added a method to disconnect GATT connection.
3. Added method to retrieve local address.

Bug: 371095923
Test: atest MultiDeviceRangingTestCases
Test: atest CtsBluetoothMultiDevicesTestCases
Change-Id: I62df16b64d5caae96dab6bc9c14f2dd8016c3ec3