Add CTS for converting java hint sessions to native hint sessions
Add an API called "APerformanceHint_getSessionFromJava" for Java
clients to access the native APerformanceHintSession objects that
Java hint sessions wrap. This should will make the Java and Native APIs
substantially more interoperable.

This patch also makes the native code aware of when it is being used in
Java, to allow it to have slightly different behavior, especially when
it is being directly accessed through JNI. The main place where this
mostly matters is in preventing clients from closing Java-owned sessions
through the JNI, as that could cause weird behavior otherwise.

Bug: 367803904
Test: atest HintManagerServiceTest
Test: atest PerformanceHintNativeTestCases
Test: atest cts/tests/tests/os/src/android/os/cts/PerformanceHintManagerTest.java

Change-Id: Iccc4ca384ee9c3f77f94913f6f6b1264c2a95733
Flag: EXEMPT test