DO NOT MERGE Improve reliability and readability of SurfaceControlTest
SurfaceControlTest has flaws in cleaning up its resources
and excessive synchronization making it difficult to read
that in the end did not provide any value. The major problem
is that result of SurfaceTransations take some time to apply
so when taking screenshot (used for validation) they are not
applied and the test failed. The best solution was to add a short
sleep after transactions have been applied.
This patch is only needed for Android12 (S)

Test: atest CtsWindowManagerDeviceTestCases
:android.server.wm.SurfaceControlTest

Bug: 379761553
Change-Id: Ibab35b433e9beddcd8dbeed2768fe1fe9a9c4e8e