[ITS][Zoom] choose shared marker closest to transition image if possible
Details: Choosing the shared marker that minimizes the average distance to the center of the images can result in significant run-to-run variation, and also makes it more difficult to diagnose smooth zoom issues. By choosing the shared marker closest to the transition image, behavior can be more consistent.

Change-Id: Ifa59f642f30a5cfc0b6a4425afd4e11b29f8af8b
Flag: TEST_ONLY
Bug: 353723916
Test: ran test_zoom and test_zoom_tele