Remove flagged class construction from test init
Using RequiresFlagEnabled at the class level only prevents the code in @Test and @Before from being executed, but objects contructed during init are still invoked. So if the object being constructed is a flagged api where the flag is disabled, the test execution will crash. So instead, we move those constructions to be on-demand so they are only invoked when needed by @Before and @Test methods.

Bug: 380151982
Flag: com.android.settingslib.flags.settings_catalyst
Test: atest CtsSettingsPreferenceServiceTest
Change-Id: I69d1d73a723a6f2271a7b7ea713568f57c96c26e