Add CheckFlagRule to DeviceConfigApiTests
A new API, getAdbWritableNamespaces, was added to DeviceConfig
and guarded by a flag. testGetAdbWritableNamespaces_returnsNamespaces
was added to DeviceConfigApiTests to verify this new API with a
RequiresFlagsEnabled annotation to only run when the flag is
enabled. However, this class does not currently have a
CheckFlagsRule, so the test was still being executed when the
flag was disabled and the API was not available. This commit adds
the CheckFlagsRule to this test class to ensure all tests annotated
with RequiresFlagsEnabled are skipped.

Fixes: 380350157
Flag: TEST_ONLY
Test: atest DeviceConfigApiTests
Change-Id: I7a309aabd28f713170b9b0aeaf3e6df2fc83a1b5