Change again how SystemApiAnnotationTest is ignored for JUnit 3
The change in
https://android-review.googlesource.com/c/platform/cts/+/3367719
didn't fix the issue. Even though the test function was made
private, the framework still tried to run the test and failed the
access check.

If I try changing the function name, then the test run fails
because there is no test to run. In the end, I've let the test
function as it is, but remove its contents. This way the test
passes.

Unfortunately I can't find another way to skip the test, or
consistently log that the test is skipped.

Bug: 378992946
Test: run cts --include-filter "CtsSystemApiAnnotationTestCases android.signature.cts.api.AnnotationTest#testAnnotation"
Merged-In: Idd4df74d17cd396a1e751279e221497889702ed4

Change-Id: I99e6c7cd1349e46e0734e60befa70cf948fec70a