WebView: refactor tests off of UI thread
This refactors three CTS test cases to use the test thread instead of
the UI thread. The thing in common between all three tests is that they
rely on the back-forward-list, which is now supported by
WebViewOnUiThread. This is a speculative change which we hope will
improve test robustness.

Bug: 372644502
Test: atest CtsWebkitTestCases:WebViewTest
Flag: TEST_ONLY
Change-Id: I580284c4d6ce05b867b84433da6e954092584f6a