Move DrmFramework tests using mpeg2 extractor to module
DrmFramework's MediaDrmClearkeyTest was marked as entirely
framework-specific; the mpeg2 tests depend on the extractor in the
module, so we move them to @ModuleSpecificTest.

This would have flagged the failure sooner because we'd have seen the
failure as part of module-specific tests instead of old-release baseline
testing.

Bug: 379803959
Flag: EXEMPT test fix
Test: CtsMediaDrmFrameworkTestCases:android.media.drmframework.cts.MediaDrmClearkeyTest
Change-Id: I983af848a39d352c68ab15bcb87d3a8fec30ff5b