Install package on current user to pass on HSUM devices
Since we set install-dependency role to current user, the package needs
to be installed on current user too. If we install for all users, then
for user 0, system fails to find a role holder. On HSUM devices, the
test runs on main user which is different than 0.

Bug: 380402322
Bug: 372862145
Test: atest PackageManagerShellCommandInstallTest
FLAG: android.content.pm.sdk_dependency_installer
Change-Id: I83e8a80215855b0b980fdd9ce67c880c801ec5fd